extract resultset package out of server package
