+++ server/BUILD.bazel
@@ -55,6 +55,7 @@ go_library(
         "//server/internal/dump",
         "//server/internal/handshake",
         "//server/internal/parse",
+        "//server/internal/resultset",
         "//server/internal/util",
         "//server/metrics",
         "//session",
@@ -185,6 +186,7 @@ go_test(
         "//server/internal/column",
         "//server/internal/handshake",
         "//server/internal/parse",
+        "//server/internal/resultset",
         "//server/internal/util",
         "//session",
         "//sessionctx",
