+++ server/internal/resultset/cursor.go
@@ -1,0 +1,55 @@
+// Copyright 2023 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package resultset
+
+import "github.com/pingcap/tidb/util/chunk"
+
+// CursorResultSet extends the `ResultSet` to provide the ability to store an iterator
+type CursorResultSet interface {
+	ResultSet
+
+	StoreRowContainerReader(reader chunk.RowContainerReader)
+	GetRowContainerReader() chunk.RowContainerReader
+}
+
+// WrapWithCursor wraps a ResultSet into a CursorResultSet
+func WrapWithCursor(rs ResultSet) CursorResultSet {
+	return &tidbCursorResultSet{
+		rs, nil,
+	}
+}
+
+var _ CursorResultSet = &tidbCursorResultSet{}
+
+type tidbCursorResultSet struct {
+	ResultSet
+
+	reader chunk.RowContainerReader
+}
+
+func (tcrs *tidbCursorResultSet) StoreRowContainerReader(reader chunk.RowContainerReader) {
+	tcrs.reader = reader
+}
+
+func (tcrs *tidbCursorResultSet) GetRowContainerReader() chunk.RowContainerReader {
+	return tcrs.reader
+}
+
+// FetchNotifier represents notifier will be called in COM_FETCH.
+type FetchNotifier interface {
+	// OnFetchReturned be called when COM_FETCH returns.
+	// it will be used in server-side cursor.
+	OnFetchReturned()
+}
